Refactor Details page sections to reuse code
