--- src/services/ui/src/components/PackageDetails/index.ts
@@ -1,1 +1,0 @@
-export * from "./ChipSpaPackageDetails";
