+++ src/services/ui/src/components/index.tsx
@@ -11,10 +11,8 @@ export * from "./ErrorAlert";
 export * from "./HowItWorks";
 export * from "./Layout";
 export * from "./LoadingSpinner";
-export * from "./PackageDetails";
 export * from "./RaiList";
 export * from "./SearchForm";
-export * from "./SubmissionInfo";
 export * from "./Modal";
 export * from "./Dialog";
 export * from "./Modal/ConfirmationModal";
