+++ src/services/ui/src/pages/detail/index.tsx
@@ -3,12 +3,10 @@ import {
   Alert,
   Attachmentslist,
   CardWithTopBorder,
-  ChipSpaPackageDetails,
   DetailsSection,
   ErrorAlert,
   LoadingSpinner,
   RaiList,
-  SubmissionInfo,
   ConfirmationModal,
 } from "@/components";
 import { useGetUser } from "@/api/useGetUser";
@@ -22,6 +20,8 @@ import { useGetPackageActions } from "@/api/useGetPackageActions";
 import { PropsWithChildren, useState } from "react";
 import { DETAILS_AND_ACTIONS_CRUMBS } from "@/pages/actions/actions-breadcrumbs";
 import { API } from "aws-amplify";
+import { DetailItemsGrid } from "@/pages/detail/components/DetailItemsGrid";
+import { spaDetails, submissionDetails } from "@/pages/detail/setup/spa";
 
 const DetailCardWrapper = ({
   title,
@@ -187,10 +187,9 @@ export const DetailsContent = ({ data }: { data?: ItemResult }) => {
           />
           <PackageActionsCard id={data._id} />
         </section>
-        <DetailsSection id="package-details" title="Package Details">
-          <ChipSpaPackageDetails {...data?._source} />
-        </DetailsSection>
-        <SubmissionInfo {...data?._source} />
+        <h2 className="text-xl font-semibold mb-2">{"Package Details"}</h2>
+        <DetailItemsGrid displayItems={spaDetails(data._source)} />
+        <DetailItemsGrid displayItems={submissionDetails(data._source)} />
         {/* Below is used for spacing. Keep it simple */}
         <div className="mb-4" />
         <DetailsSection id="attachments" title="Attachments">
